+++ meteor-app/imports/ui/ordering/cart-list-item.stories.js
