+++ meteor-app/imports/ui/ordering/cart.js
@@ -1,15 +1,20 @@
 import React, { Fragment } from 'react'
 import { withRouter } from 'react-router-dom'
 import { Grid } from 'semantic-ui-react'
+import CartList from '/imports/ui/ordering/cart-list'
 
 class Cart extends React.Component {
-
   render() {
     return (
       <Grid>
-        <Grid.Row>  
-          <Grid.Column width='13'>
-            <div><h1>CART / ORDER FULL OF PARTS LISTED HERE </h1></div>
+        <Grid.Row>
+          <Grid.Column>
+            <CartList
+              title={'Cart List:'}
+              order={this.props.order}
+              componentClassName='cart-list-main'
+              loading={this.props.loading}
+            />
           </Grid.Column>
         </Grid.Row>
       </Grid>
